requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pymysql>=1.0.3
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
pydantic>=1.10.7
requests>=2.28.2